requirements.txt
mcp[cli]>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
